# contracts/openapi/v1/openapi.json
{
  "components": {
    "schemas": {
...
      "ClassifyInboxRequest": {
        "description": "Inbox 아이템을 프로젝트에 연결 확정.",
        "properties": {
          "projectIds": {
            "items": {
              "type": "string"
            },
            "title": "Projectids",
            "type": "array"
          }
        },
        "required": [
          "projectIds"
        ],
        "title": "ClassifyInboxRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InboxPromoteIn": {
        "description": "POST /inbox/{id}/promote 요청 — 대상 team workspace.",
        "properties": {
          "targetWorkspaceId": {
            "format": "uuid",
            "title": "Targetworkspaceid",
            "type": "string"
          }
        },
        "required": [
          "targetWorkspaceId"
        ],
        "title": "InboxPromoteIn",
        "type": "object"
      },
      "InboxPromoteOut": {
        "description": "POST /inbox/{id}/promote 응답 — 복제본 + audit 식별자.\n\nInboxItem 은 source_type='inbox' EmbeddingChunk 가 존재하지 않으므로\nBG embedding 복제 없이 status='completed' 즉시 반환 (notes/meetings 는\n'embedding_pending' — BG 흐름 차이).",
        "properties": {
          "audit_id": {
            "format": "uuid",
            "title": "Audit Id",
            "type": "string"
          },
          "new_inbox_id": {
            "format": "uuid",
            "title": "New Inbox Id",
            "type": "string"
          },
          "status": {
            "default": "completed",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "new_inbox_id",
          "audit_id"
        ],
        "title": "InboxPromoteOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Kairos API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/inbox": {
      "get": {
        "operationId": "list_inbox_api_v1_workspaces__workspace_id__inbox_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isProcessed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Isprocessed"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Inbox",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/inbox/{inbox_id}/classify": {
      "post": {
        "operationId": "classify_inbox_api_v1_workspaces__workspace_id__inbox__inbox_id__classify_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "inbox_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inbox Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyInboxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify Inbox",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/inbox/{inbox_id}/dismiss": {
      "post": {
        "operationId": "dismiss_inbox_api_v1_workspaces__workspace_id__inbox__inbox_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "inbox_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inbox Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dismiss Inbox",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/inbox/{inbox_id}/promote": {
      "post": {
        "description": "Inbox 아이템 → team workspace 복제 + audit row.\n\n202 Accepted — InboxItem 은 source_type='inbox' EmbeddingChunk 가 실제 인서트되지 않음\n(whitelist 만 존재) → BG embedding 복제 없음. ItemPromotionAudit.embedding_status='n/a'.\nai_suggested_project_id 는 복제본에서 None (cross-workspace 제약 — composite FK).",
        "operationId": "promote_inbox_api_v1_workspaces__workspace_id__inbox__inbox_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "inbox_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inbox Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxPromoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxPromoteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote Inbox",
        "tags": [
          "inbox"
        ]
      }
    },
...
  }
}